+++ src/coreclr/System.Private.CoreLib/src/System/MulticastDelegate.CoreCLR.cs
@@ -14,28 +14,13 @@ namespace System
 {
     [ClassInterface(ClassInterfaceType.None)]
     [ComVisible(true)]
-    public abstract class MulticastDelegate : Delegate
+    public abstract partial class MulticastDelegate : Delegate
     {
         // This is set under 2 circumstances
         // 1. Multicast delegate
         // 2. Wrapper delegate
         private object? _invocationList; // Initialized by VM as needed
         private nint _invocationCount;
-
-        // This constructor is called from the class generated by the
-        //    compiler generated code (This must match the constructor
-        //    in Delegate
-        [RequiresUnreferencedCode("The target method might be removed")]
-        protected MulticastDelegate(object target, string method) : base(target, method)
-        {
-        }
-
-        // This constructor is called from a class to generate a
-        // delegate based upon a static method name and the Type object
-        // for the class defining the method.
-        protected MulticastDelegate([DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)] Type target, string method) : base(target, method)
-        {
-        }
 
         internal bool IsUnmanagedFunctionPtr()
         {
@@ -446,34 +431,6 @@ namespace System
             {
                 return ((uint)index < (uint)_invocationCount) ? invocationList[index] : null;
             }
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool operator ==(MulticastDelegate? d1, MulticastDelegate? d2)
-        {
-            // Test d2 first to allow branch elimination when inlined for null checks (== null)
-            // so it can become a simple test
-            if (d2 is null)
-            {
-                return d1 is null;
-            }
-
-            return ReferenceEquals(d2, d1) ? true : d2.Equals((object?)d1);
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool operator !=(MulticastDelegate? d1, MulticastDelegate? d2)
-        {
-            // Can't call the == operator as it will call object==
-
-            // Test d2 first to allow branch elimination when inlined for not null checks (!= null)
-            // so it can become a simple test
-            if (d2 is null)
-            {
-                return d1 is not null;
-            }
-
-            return ReferenceEquals(d2, d1) ? false : !d2.Equals(d1);
         }
 
         public sealed override int GetHashCode()
