+++ src/mono/System.Private.CoreLib/src/System/MulticastDelegate.Mono.cs
@@ -10,20 +10,9 @@ using System.Runtime.Serialization;
 namespace System
 {
     [StructLayout(LayoutKind.Sequential)]
-    public abstract class MulticastDelegate : Delegate
+    public abstract partial class MulticastDelegate : Delegate
     {
         private Delegate[]? delegates;
-
-        [RequiresUnreferencedCode("The target method might be removed")]
-        protected MulticastDelegate(object target, string method)
-            : base(target, method)
-        {
-        }
-
-        protected MulticastDelegate([DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)] Type target, string method)
-            : base(target, method)
-        {
-        }
 
         [Obsolete(Obsoletions.LegacyFormatterImplMessage, DiagnosticId = Obsoletions.LegacyFormatterImplDiagId, UrlFormat = Obsoletions.SharedUrlFormat)]
         [EditorBrowsable(EditorBrowsableState.Never)]
@@ -273,22 +262,6 @@ namespace System
 
                 return ret;
             }
-        }
-
-        public static bool operator ==(MulticastDelegate? d1, MulticastDelegate? d2)
-        {
-            if (d1 == null)
-                return d2 == null;
-
-            return d1.Equals(d2);
-        }
-
-        public static bool operator !=(MulticastDelegate? d1, MulticastDelegate? d2)
-        {
-            if (d1 == null)
-                return d2 != null;
-
-            return !d1.Equals(d2);
         }
 
         internal override object? GetTarget()
